+++ src/interfaces/IRecursoDigital.java
@@ -4,9 +4,6 @@ public interface IRecursoDigital {
         String getTitulo();
         String getAutor();
         int getAnioPublicacion();
-        boolean estaDisponible();
-        void marcarComoPrestado();
-        void marcarComoDisponible();
 }
