+++ src/interfaces/Prestable.java
@@ -1,0 +1,7 @@
+package src.interfaces;
+
+public interface Prestable {
+    boolean estaDisponible();
+    void marcarComoPrestado();
+    void marcarComoDisponible();
+}
