+++ src/modelos/RecursoDigital.java
@@ -31,21 +31,6 @@ public abstract class RecursoDigital implements IRecursoDigital {
         return this.anioPublicacion;
     }
 
-    @Override
-    public void marcarComoPrestado() {
-        disponible = false;
-    }
-
-    @Override
-    public void marcarComoDisponible() {
-        disponible = true;
-    }
-
-    @Override
-    public boolean estaDisponible(){
-        return this.disponible;
-    }
-
     public void setTitulo(String titulo){
         if (titulo == null || titulo.trim().isEmpty()){
             throw new IllegalArgumentException("El titulo no puede estar vacio");
